+++ Contents/mods/B42 PZLinux/42/media/lua/client/Context/World/Features/PZLinuxRequest.lua
@@ -71,6 +71,7 @@ function requestUI:initialise()
 
     function self.topBar.onMouseDown(topBar, _x, _y)
         topBar.parent.isDragging = true
+        PZLinuxTrackDragging(topBar.parent)
         topBar.parent.initialX = topBar.parent:getX()
         topBar.parent.initialY = topBar.parent:getY()
         topBar.parent.mouseStartX = getMouseX()