+++ Contents/mods/B42 PZLinux/42/media/lua/client/Context/World/Features/PZLinuxTrading.lua
@@ -88,6 +88,7 @@ function tradingUI:initialise()
 
     function self.topBar:onMouseDown(_x, _y)
         self.parent.isDragging = true
+        PZLinuxTrackDragging(self.parent)
         self.parent.initialX = self.parent:getX()
         self.parent.initialY = self.parent:getY()
         self.parent.mouseStartX = getMouseX()